Bump Go version to 1.16
This allows Go to build a native binary for Apple Silicon.

GoReleaser is used to build and release this project, GoReleaser
v0.157.0 released on 2021-02-17 will build a fitting binary
automatically as long as Go 1.16 is used. A bump of the Go version
should be everything that's needed to get this working natively on Apple
Silicon.